Introducing digital signatures for SEPA Direct Debits
This introduces a digital signing box for the new member subscription
page. When users enter their IBAN, they must now also draw their
signature.

The signature is drawn by the user on an HTML canvas. Its data is
submitted with the whole form. On the server, the string is converted
into an actual image and stored to be accessible later.