Correctly implement canvas resizing to prevent crashes
This change is introduced to stop the subscription webpage from crashing
on mobile devices. This issue was caused by implementing the canvas
resizing using the `window.onresize` EventHandler, however, on mobile
devices this event is also emitted when scrolling (due to the moving
omnibox).

Resizing is now done when the orientation of a mobile device changes.

Furthermore, the height of the canvas is increased from 100 to 150 to
aid mobile users entering their signature.