Update membership types
This updates the membership types that are available to members in
accordance with the bylaws from July 01, 2021. The association
knows the following membership types for members in the sense of
the law:

- gewoon (ordinary)
- extern (external)
- erelid (honorary)

An additional membership type "afgestudeerde" (graduate) is
available to those who were previously a member (as defined above).