Change how the expiration of a member's membership is determined
`Member`s now have a new attribute `membershipEndsOn`, this
attribute is required together with the upcoming checks against
the TU/e student administration database.

The attribute can be either `null` or a date. If it is `null`, the
membership of the user has a rolling expiration date. In other
words, it will always be the end of the current association year.
If `membershipEndsOn` is an actual date, the membership of that
member will have ended on that specific date. The date may be in
the future, in which case the member's current membership type is
still valid.

Changing the type of membership of a `Member` correctly updates
the associated expiration of said membership.