Add checks for membership type and status
This adds functionality to the `Checker` module that enables automatic
lookups of members against the TU/e student administration database.

The idea is that this runs every night and it will automatically select
a group of members that will have their membership status checked. This
means that if the member is still studying at the M&CS department,
nothing changes. However, if they are no longer studying at the
department, their membership will get an end date (`membershipEndsOn`).

Eventually, this date is reached and at that point their membership is
converted into `graduate` or `external`. The latter being the case if
the member is a so-called "active member" (i.e., they are a member of
an organ).