Determine membership expiration and generation on approval of prospective member
Since the membership expiration and generation depend on when the
member is approved, and not when they 'subscribed' this has been
moved from the `ProspectiveMember` to just `Member`.

This also ensures that the generation of a member is always the first
year of the current association year (e.g, on May 17, 2022, the
generation will be 2021 instead of 2022).