Fix ordering of `null` values in list of members to check
Postgres sorts `null` values after anything else in `ASC` order. This
is not what we want to happen, as we like to check members who have
never been checked before everyone else.

This is supported by Postgres through `NULLS FIRST`, however, Doctrine
does support this in ORM. Hence, we need to do it differently. While
implementing a custom `SqlWalker` is the cleanest solution, the
current solution suffices for quite some time.

An additional hidden column is added to the query, which depending
on whether `lastCheckedOn` is `null` will either be 0 or 1
(respectively) which is used as an initial sort, after that the
normal sort is used.