Upgrade to ZF3
This upgrades the Zend Framework to v3, which includes significant
changes to the service manager. The `ServiceManagerAwareInterface`
has been removed, making dependency injection necessary to be
able to use methods and data from other files.

The global event listener is currently disabled due to issues
with properly creating and sharing an `EventManager`. The
`SharedEventManager` looks promising but I have not been able
to get it to work.

Additionally, forms with collections appear to be broken (e.g.,
foundation of organs). There is no clear indication as to why
this does not work.