Remove custom event logger
The custom event logger was introduced to keep track of all
changes to the database. The reasoning was that it would allow
recovery if something went wrong. Unfortunately, the triggers
were often inforrectly applied meaning that you would not be
able to actually see the changed data.

Additionally, careful analysis of the database shows that the
data is also truncated, hence it cannot even be used.