Switch to native attribute for Doctrine
This also moves some things to native enums to prevent repeating
certain values a billion times. This does require additional
hydrator strategies that know how to work the values.

Additionally, an increased number of attributess in the models
has a default value (probably an empty string) and certain
attributes are not strongly typed. This is because of how the
hydrators work (i.e., a design flaw).