Fix hydration issues due to typed properties
This fixes the issues that presented itself after converting all
entity properties to be typed. The `DoctrineObject` hydrator
allows properties to be extracted/populated by reference instead
of the actual entity API (i.e., the getter and setter methods).

Unfortunately, this partially breaks the `ProspectiveMember`
approval form, as addresses disappear.