Fix organ foundation
It is now once again possible to create new organs. I have not been
able to figure out what was the actual underlying problem (perhaps
a bug in `laminas/laminas-form`?), so this solution is a replica of
what is supposed to happen.

Instead of binding a new decision to the form, we simply replicate
the behaviour by manually using the foundation hydrator. This
required some changes due to how the properties of the different
entities are typed.

Additionally, this required some changes to the `Report` module,
which is really hacky by using reflection. But it works, so I'll
leave it at that.