Make membership type changes retroactive again
This puts back the original idea of how manual membership type
changes are computed. They are now retroactively applied. Any
changes made will appear to have been like this since the end of
the previous association year.

For example, if a member is currently `graduate` on 2022-08-04
with a `membershipEndsOn` of 2022-07-1, changing them to `member`
removes the `membershipEndsOn` and puts their `expiration` on
2023-07-01.

Another example, if a member is currently `member` on 2022-08-04,
changing them to `external` will put `membershipEndsOn` to
2022-07-01 and their `expiration` on 2023-07-01.