Update dependencies and fix language switching enrolment form
Updates the dependencies because I am stupid, but they needed to
be updated anyway.

Fixes the redirect when changing the language form. As the
language can only be changed on the enrolment form, if the user is
not logged in, they were likely coming from the enrolment form.

Furthermore, the enrolment form is now in English by default.