Update `Checker` to search for issues with "Inactief Lid" of organs
The `Checker` can now determine that a member cannot be "Lid" AND
"Inactief Lid" of the same organ, it guesses whether or not the
member should actually be "Lid" or "Inactief Lid", however, there
is no guarantee this is correct.

Additionally, it can detect "Inactief Lid" members who still have
a special role in the organ.

Co-Authored-By: rinkp <rink@pietershgv.nl>