Fix typing for Psalm to resolve polymorphic calls
This is the first time that Psalm actually kept complaining about
polymorphic calls. This has been address by making the `Checker`'s
`Error` a template for the generic type `T`. All specific errors
that extend the class now also extend it with a specific class for
implementation.

Also removes the budget check, as that did not do anything.