Fix issue where installing/discharging members from an organ was not possible
The check to determine if a meeting and related subdecision exist
when installing/discharging members in/from organs was too strict.
It required that the (current) meeting and the foundation sub-
decision were from the same meeting. This is not desirable, as
changing to organs are almost never done in the same meeting.