Add `deleted` state to `Member`
Determines if a member is deleted. A deleted member is a member
whose basic info needs to be retained to ensure that all decisions
that mention this member can be kept (i.e., administrative
purposes). This value is only set when deleting a member and
cannot be altered via the interface.