Add member updates
This adds the basic infrastructure that will support data update
requests from members from the GEWIS website.

This still requires an ingress point for the requests (including
a way to authenticate a member from the website) and a webhook
that will allow us to automatically update the data (after
approval) on the website.