Fix Doctrine event listeners not working when using CLI
When using `laminas/laminas-cli` the MVC-part of the Application is
never used. This means that the `MvcEvent` needed for all the
`onBootstrap()`s is never called, hence, the event listeners would
not be registered (as they were part of the `Database` module's
`onBootstrap()`).

This has been fixed by moving the registration of the event listeners
to a place that is always executed by Laminas' `ModuleManager`,
regardless of whether we use the CLI or the web interface.

Usage of `loadModules.post` is necessary to ensure that all modules
are loaded before registering the listeners.