Fix Doctrine listeners not loading due to too late config override
When running tests through PHPUnit, we override the config of Doctrine
to force it to use an in-memory SQLite database. Unfortunately, this
override happens after `loadModules()` is called, which means that the
event listeners have already been registered before the actual config
has been changed.

This caused Doctrine's `EntityManager` in the listeners to use the old
connection parameters (i.e. a PostgreSQL database that dix not exist).

Because the tests use the production configs I have added a new `test`
configuration, which takes all the production configs but directly
overrides the necessary values instead of doing this during the setup
phase of the application in PHPUnit.

---

I initially tried to delay the loading of our own modules, however,
this proved to be impossible, as Laminas needs to know which modules
can be loaded. These cannot be added after the fact.

Changing how the Doctrine listeners are registered is also not an
option, as that would probably mean that it broke the CLI again.

---

Thanks to @rinkp for figuring this out.