Check for staged files and not unstaged files
If the `git add --all` works, all files are staged. This means that `git diff --exit-code` will return `0` (there are no changed files). However, this is problematic, as this prevents us from actually committing stuff we want to commit.

To fix this, we have to use `git diff --quiet --cached` to determine if there are staged changes or not (note: `--quiet` is needed to prevent output if there are differences).