Don't show historic organ installations for simple members
Due to my refactoring w.r.t. the retrieval of `Member` objects the
installations shown on a member's information page would be incorrect
if the member was retrieved as a "simple" member.

These members do not have any "active" installations, but they can
have old installations. This means that by calling
`getInstallations()` all these installations would be returned. This
should not be the case.

Co-Authored-By: Rink Pieters <rinkp@users.noreply.github.com>