Fix e-mails not being sent when headers contain unicode
Unfortunately, the e-mail protocols are still archaic. Even with
RFCs like 6532, there is no support for actual unicode characters
in e-mail headers.

We can circumvent this to a certain degree by re-encoding strings
that may contain unicode characters. This is done through
`mb_encode_mimeheader`. However, it should be pointed out that this
function has a hard line length of 74 characters. This is rather
annoying (and incredibly stupid), as it also breaks the e-mail
headers (CRLF injection), so we also have to fix this by replacing
the EOL with nothing.