Add types to the view templates
This also introduces a `HelperTrait` which contains the view definitions
of all the relevant Laminas packages. On top of that it also contains
the definitions of our own view helpers.

I have not added types to everything, all variables that are accessed
through a magic method have not been touched.