Remove everything related to IBANs
With the upcoming payment flow integration in the enrolment expierence
we do no longer need to keep track of IBANs or anything related to it.

The form, its validation, the models, and supporting code have all been
stripped of relavent parts. This also removes the JavaScript library
used to obtain the signatures necessary for SEPA Direct Debit.

This also enables the enrolment form again after having been disabled
in GH-303.