Add (basic) Stripe payment flow
It ended up being a bit less basic than I had hoped. In short, it adds
a payment flow to the enrolment/registration form.

The following steps are performed:
- prospective member submits the form
- if correct, the prospective member is persisted
- a payment link is created and persisted*
- an e-mail is sent with details about the registration
- a Checkout Session for Stripe is created
- prospective member is redirected to the Stripe checkout page

The prospective member is requested to pay:
- if cancelled, the prospective member returns to the website
- if successful, we need to wait on the Stripe webhooks to fire

We listen for four specific webhooks:
- checkout.session.expired: if the prospective member did not complete
  the checkout within 24 hours, we get notified. We can use this event
  to send a reminder e-mail to the prospective member.
- checkout.session.completed: when the prospective member successfully
  completed the checkout, however, this does not mean that the payment
  is finished.
- checkout.session.async_payment_succeeded: delayed payment finally
  succeeds.
- checkout.session.async_payment_failed: delayed payment failed.

*) The payment link is meant for (re)starting a checkout session, this
may be necessary when the prospective member has (accidentally)
cancelled the checkout, forgot to checkout within 24 hours, or after a
payment failure.

The checks to make sure that everything is aligned are quite extensive
but there are some assumptions on how things are expected to work.