Change how we display prospective member related information
To align the prospective member overviews with when they can be
approved or deleted the columns have changed from

- created, expired, failed
- pending
- paid

to

- created, pending
- expired, failed
- paid

Additionally, the Checkout Session ID is now also used to display
a "personal" message on the checkout cancelled/completed page.

Contains a minor fix for determining the expiration of a Checkout
Session when trying to restart it and a fix for the layout on the
checkout status pages, which used the wrong navbar.

Co-Authored-By: Bartjan Henkemans <38168308+BHenkemans@users.noreply.github.com>