Handle manual approval of prospective members and potential refunds
This changes the approval and removal flow for prospective members.

A prospective member can now be approved if their last Checkout
Session state is `EXPIRED` or `FAILED`. This requires unfolding the
approval form (adding an additional checkmark it too much work).

---

Prospective members can now only be deleted if their last Checkout
Session `FAILED`, when they are fully `EXPIRED` or when they have
paid (`PAID`).

In the event that they have paid, we will try to create a Refund
for the original Charge. If this fails, the prospective member is
not removed and the secretary is warned about this.

If the Refund is created successfully Stripe will handle everything
else, it may take up to 5-10 business days before the Refund is
fully processed by Stripe.

To make this possible, when we mark a Checkout Session as `PAID` we
store the id of the Payment Intent, which means we have to make one
API call less when we are trying to create a Refund. Because Refunds
work with a specific Charge and not a Checkout Session or Payment
Intent.

If Stripe fails to complete the Refund request, we handle the
`charge.refund.updated` event for the following states: `failed`,
`requires_action` or `canceled`. The secretary will receive an
e-mail for this, as the refund should potentially be done in another
way.