Fix removal query to find fully expired Checkout Sessions
Since fixing the processing of the expiration of a restarted Checkout
Session, prospective members were deleted too soon. This was the
result of the latest Checkout Session always having an expiration of
a day (ONLY IF RESTARTED).

This fixes that by adjusting the query to make sure that if we are
dealing with an expired and restarted Checkout Session we always
check the `recoveredFrom` (i.e. original) Checkout Session to check
whether the limit has passed.