Fix restarting Checkout Sessions and handling of events after paying
The `PENDING` or `PAID` check has been move before trying to get the
original Checkout Session. This should prevent the prospective member
from restarting/recovering the Checkout Session if the checkout link
is accidentally activated.

Additionally, when handling an event that is part of a recovery we
do not process any events if the status of the last know (recovered)
Checkout Session is `PAID`.