Add `Reappointment` subdecision for installations in bodies
This adds a new type of organ mutation, a reappointment. This mutation
can be used to "prolong" the membership of a member in an organ. The
subdecision refers to the original installation, just like a discharge.

While a self-reference would have been nice (for consecutive
reappointments), it introduces a lot of compexity due to the added
recursion. As such, `Reappointment`s have a many-to-one relationship
with `Installation`s.

`Reappointment`s are processed before `Discharge`s and `Installation`s.

In the Organ Mutations UI, only the `Lid` function can be directly
reappointed. All other functions can be reappointed by checking the
relevant functions in the model. This means that if a function is NOT
checked it will not be reappointed, this means that a discharge for
this particular function is created.

Some of the JavaScript discharge functionality has been moved into a
separate function to reduce code duplication. On top of that, the
reappointments, discharges, and installations are now also sorted as
such in the UI.