Move any `Member` association to `SubDecision` itself
Instead of allowing specific types of sub-decisions to specify an
association with `Member` this is now done "globally". If there
are specific needs (e.g. not nullable) this can be achieved using
`AssociationOverride`s. Note: in the `JoinColumn` override it is
necessary to (re)specify `name` and `referencedColumnName`.

This change is necessary to allow easier extraction of sub-
decisions related to specific members (i.e. for GDPR requests).

Functionally nothing changes as all the different associations
already super-mapped to the same column.