Make it clear `OrganRegulation`s are only for committees and fraternities
The other organ types are still shown on the form, however, the
options are disabled. This should prevent people from misusing the
existing options for other things.

If the user decides to edit the form to forcefully use another organ
type this is prevented through form validation. In the event even
this fails to catch the mistake there is a hard fail-safe in the
`OrganRegulation` hydrator.