Add non-official English translations to `SubDecision`s
`getContent()` remains the same, however, an aptly named
`getAlternativeContent()` is now also required. Both functions use
their respective template (`getTemplate()` and
`getAlternativeTemplate()`) and perform replacements on them using
the `replaceContentPlaceholders()` in `SubDecision`.

This change does introduce some duplication, however, creating a
whole new system is not a good idea. Furthermore, putting the
templates or parts of it in the translation files is also not
foolproof. Accidental changes to translations should NEVER lead to
changes in the content of a (sub)decision.

The alternative content is never stored in the database and thus
also not propagated to ReportDB (i.e. GEWISWEB).

The "Copy Decision"-button has been updated to automatically include
the alternative content next to the original content formatted as
the LaTeX `\decision` command.

Ideally, there would have been tests to ensure that replacement tags
present in each of the four functions defined in each subdecision.
However, this takes a long time to implement. An alternative would
be to use PHPUnit and create unit tests that can achieve this. Such
a unit test may look as follows:

```php
public function testReplacementsInSpecificSubDecision(): void
{
    $subDecision = new SpecificSubDecision();

    $this->assertStringContainsString(
        '%REPLACEMENT_TAG%',
        $subDecision->getTemplate(),
    );
    $this->assertStringContainsString(
        '%REPLACEMENT_TAG%',
        $subDecision->getAlternativeTemplate(),
    );

    $replacements = $subDecision->getReplacements();
    $this->assertArrayHasKey('%REPLACEMENT_TAG%', $replacements);

    $alternativeReplacements =
        $subDecision->getAlternativeReplacements();
    $this->assertArrayHasKey(
        '%REPLACEMENT_TAG%',
        $alternativeReplacements,
    );
}
```