Ensure that LaTeX special characters are escaped
Adds functionality to enable the escaping of special characters in
LaTeX.

The ordering of the replacements is of utmost importance to prevent
creating illegal LaTeX commands or mangling the intended output. As
such, we cannot use `str_replace()` which will replace earlier
replacements and have to use a regex to actually achieve this.