Fix copying of decisions
Due to the previous usage of `escapeHtmlAttr` extraneous characters
were added that bloated the copied string. This fixes that by moving
the decision's contents out of an attribute and into a `<span>`.

Requires `.trim()` to remove the unnecessary white-space that is
created due to our templating engine.

Also changed the colours of the copy buttons to `info` (i.e. soft
cyan).