Change `SubDecision` `number` property to `sequence`
A problem arises when a `SubDecision` wants to refer to a
`Meeting`. There is overlap in the `number` attribute with
`SubDecision`.

That is because we always had the following structure; `Meeting`
is defined by its:
- `type`
- `number`.

A `Decision` is:
- `meeting_type` (references `Meeting`.`type`)
- `meeting_number` (references `Meeting`.`number`)
- `point`
- `number`.

A `SubDecision` is:
- `meeting_type` (references `Decision`.`meeting_type`)
- `meeting_number` (references `Decision`.`meeting_number`)
- `decision_point` (references `Decision`.`point`)
- `decision_number` (references `Decision`.`number`)
- `number`

With these changes we get that a `SubDecision` is:
- `meeting_type` (references `Decision`.`meeting_type`)
- `meeting_number` (references `Decision`.`meeting_number`)
- `decision_point` (references `Decision`.`point`)
- `decision_number` (references `Decision`.`number`)
- `sequence`

While `sequence` is a non-reserved keyword in PostgreSQL, it does
not cause any issues in this context.

This change also requires a change in GEWISWEB.