Add `NotificationCount` VH to `HelperTrait` for autocomplete and change DI
The `HelperTrait` is used in most template files to annotate `$this`
which subsequently allows auto-complete to work. The function
`getNotificationCount()` is now added to that trait for auto-
completion.

Furthermore, we do not need to inject the whole Psr container into
the view helper, as such we now only inject the `FrontPage` service.