Fix test configuration for unit testing
Unlike with GEWISWEB, we previously used special `test` keyed
configuration files to load in the test configuration. This is
necessary to replace PostgreSQL with SQLite when running the unit
tests.

However, when running the unit tests, they would start failing due
to there not being any actual database. Unfortunately, we could not
just do a `./orm orm:schema-tool:update --complete --force`
in the action or locally to create the necessary schemas. The
reason for this is that the application bootstrapping for unit
testing and the CLI are done separately.

Fixing this would require us to change the `./orm` executable to
be able to distinguish between testing and normal operations. This
is cumbersome, as such we use the GEWISWEB approach and create the
schemas during the application bootstrap.

Unfortunately, we listen to `loadModules.post` to be able to add
the Database -> Report event listeners. This is necessary, because
`onBootstrap` cannot be called from the CLI. Further reasoning and
information is available in GH-216. In this event loop we load the
`orm_default` `EntityManager`, this is problematic because once we
want to override the existing application `config`, this will not
be picked up by the `ServiceManager` as `orm_default` is already
cached.

We cannot use `setServer('....orm_default', null)` to forcefully
clear the `EntityManager` from the service cache, as such we need
to listen to `EVENT_MERGE_CONFIG` while constructing the actual
`ServiceManager`. At that point, we take the freshly generated
("merged") config and combine it with our custom config to
override the required parameters.

All these changes make management of the unit test bootstrapping
and running the actual unit tests easier.

---

We previously also use a service alias `database_doctrine_em` for
`orm_default`. It appears that this is no longer necessary, as
such, all instances have been replaced with the default
`doctrine.entitymanager.orm_default`.