Fix `membershipEndsOn` not updating when making `Member` `external`
Original code was added in 0b26dfb to address an issue with the membership
change, however, later in 9f75598de the `+ 2` / `+ 1` for the expiration were
reverted to make all membership changes retroactive again. Unfortunately, the
case for `external` was never changed to reflect this. As such, the
`membershipEndsOn` was always in the past, while for `external` this is not
allowed (should always be at the end of the current assocation year).