feat: no hard-coded secretary for renewal emails
I am against having the names hard-coded in the e-mails. We previously discussed
in GH-286 that we do not have a better way of doing this, since then we have the
new dynamic configuration available that would solve this (GH-390) but would
still require updates each year to ensure the data is up-to-date.

Instead, we do a bit less personalisation and a bit more of not hard-coding
names and dates.