feat: add DBAL middleware to ensure proper role is used for all queries
Initially, we attempted to use the `wrapperClass` option to handle setting the
database role. However, we encountered issues because DBAL checks for the exact
`Connection` class type and not the interface, making it difficult to extend the
`Connection` class as needed.

We also considered overwriting the PgSQL-specific connection class (PDO variant)
to set the role upon connection. Unfortunately, this was not an option because
the class is declared as `final`, preventing us from extending it.

The next potential solution was to use an `EventSubscriber` to set the role
after the connection was established (using `postConnect`). However, this
approach is already deprecated in our version of DBAL and completely removed in
the next major release, rendering it unsuitable for us (maintainability).

Ultimately, we implemented the `SET ROLE` functionality using DBAL's middleware.
By wrapping the driver, and manually creating the `Connection` we can perform
the `SET ROLE` query before the connection is used by the application.

Runtime checks exist to ensure that the role for each database is defined.
However, validation of the actual value is done by PostgreSQL itself (it will
complain if the role does not exist).

Co-Authored-By: Rink <rinkp@users.noreply.github.com>