fix(config): WAF proxy IP addresses
Unfortunately, it was obviously not one or just 2 IP addresses that the WAF
uses, but entire ranges are reserved (with no clear overview of which IP
addresses are used in them).

Therefore now support for ranges with CIDR notation when using proxies.

Copied from https://github.com/GEWIS/gewisweb/pull/1979.