fix(decision): statutory content on export
Due to Laminas being Laminas, the `DummyTranslator` uses the deprecated variant
of the `TranslatorInterface`. This is not compatible with how we do the
translations of decision content.

This has been fixed by wrapping the `DummyTranslator` in an `MvcTranslator`
object (Laminas also does this: https://github.com/laminas/laminas-mvc-i18n/blob/433e71e/src/TranslatorFactory.php#L79).

However, the used translator then uses the source text of the installation
functions in the decision content. These were in English, resulting in Dutch
text with English functions.