fix: association overrides for member property on installation decisions
Since version `2.15.x`, ORM has not supported our association overrides for the
normal `Installation` `SubDecision`s. Currently, this only results in a
deprecation warning, however, in version `3.x` and up, this is hard error that
prevents ORM from functioning.

As a result, GEWISWEB-ng cannot function as it requires ORM version `3.5.x`.

---

To fix this, I have made the `$member` association a trait (`MemberAwareTrait`),
which is imported only for `Installation` and `Board\Installation`. For the
`Board\Installation`, the `AssociationOverride` has been removed in DB and been
simplified in ReportDB.

For `Installation`, the operation is a bit more "complex" and we lose some
functionality. We need the `inversedBy`, but we can no longer do:

```php
joinColumns: new JoinColumn(
    name: 'lidnr',
    referencedColumnName: 'lidnr',
    nullable: false,
),
```

to make the attribute not `NULL`able as this will cause ORM to incorrectly
think that the _whole_ column should be `NOT NULL` which does not work for the
other `SubDecision`s that have `$member`.